requirements.txt
kafka-python
scikit-learn
tensorflow
langchain
langgraph
streamlit
pandas
pydantic
numpy
joblib
